drivers: w1 introduce 1-wire api and common code
This commit introduces a new api for the Dallas 1-wire protocol.
The api includes link functions for read and write operations on
bit, byte, and block level, as well as
function to reset and lock the bus.

The bus configuration is derived from the device tree and can be
queried using w1_peripheral_count routine.

Instead of master/slave terminology controller/peripheral terms are
used to refer to bus participants.

Signed-off-by: Thomas Stranger <thomas.stranger@outlook.com>